lms_backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
python-decouple
python-dotenv
Pillow
django-filter
sqlparse
asgiref
tzdata
